+++ core/common/src/commonMain/kotlin/com/linkit/company/core/common/architecture/Intent.kt
@@ -1,0 +1,3 @@
+package com.linkit.company.core.common.architecture
+
+interface Intent
