+++ core/common/src/commonMain/kotlin/com/linkit/company/core/common/architecture/SideEffect.kt
@@ -1,0 +1,3 @@
+package com.linkit.company.core.common.architecture
+
+interface SideEffect
