--- core/common/src/commonMain/kotlin/com/linkit/company/core/common/architecture/sideeffect/SideEffect.kt
@@ -1,3 +1,0 @@
-package com.linkit.company.core.common.architecture.sideeffect
-
-interface SideEffect
