--- core/common/src/commonMain/kotlin/com/linkit/company/core/common/architecture/state/UiState.kt
@@ -1,3 +1,0 @@
-package com.linkit.company.core.common.architecture.state
-
-interface UiState
