+++ feature/home/src/commonMain/kotlin/com/linkit/company/feature/home/sample/HomeIntent.kt
@@ -1,6 +1,6 @@
 package com.linkit.company.feature.home.sample
 
-import com.linkit.company.core.common.architecture.intent.Intent
+import com.linkit.company.core.common.architecture.Intent
 
 sealed interface HomeIntent : Intent {
 }
