+++ feature/home/src/commonMain/kotlin/com/linkit/company/feature/home/sample/HomeSideEffect.kt
@@ -1,6 +1,6 @@
 package com.linkit.company.feature.home.sample
 
-import com.linkit.company.core.common.architecture.sideeffect.SideEffect
+import com.linkit.company.core.common.architecture.SideEffect
 
 sealed interface HomeSideEffect : SideEffect {
 }
