+++ feature/home/src/commonMain/kotlin/com/linkit/company/feature/home/sample/HomeUiState.kt
@@ -1,6 +1,6 @@
 package com.linkit.company.feature.home.sample
 
-import com.linkit.company.core.common.architecture.state.UiState
+import com.linkit.company.core.common.architecture.UiState
 
 data class HomeUiState(
     val isLoading: Boolean,
